requirements.txt
numpy
pandas
scipy
pyyaml
requests
urllib3<2
plotly
streamlit
